requirement.txt
-i https://pypi.org/simple
backports.csv==1.0.7
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cheroot==6.5.4
cherrypy==18.1.1
feedparser==5.2.1
flask==0.10.1
future==0.17.1
idna==2.8
itsdangerous==1.1.0
jaraco.functools==2.0
jinja2==2.10.1
lxml==4.3.3
markupsafe==1.1.1
more-itertools==7.0.0
mysqlclient==1.4.2.post1
nltk==3.4.1
numpy==1.16.2
pattern==3.6.0
pdfminer.six==20181108
portend==2.4
pycryptodome==3.8.1
pyspellchecker==0.4.0
python-docx==0.8.10
pytz==2019.1
requests==2.21.0
scipy==1.2.1
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.1
tempora==1.14.1
urllib3==1.24.2
werkzeug==0.15.2
zc.lockfile==1.4
